requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
faiss-cpu==1.8.0
numpy==1.26.4
httpx==0.27.0
sentence-transformers==3.0.1
python-dotenv==1.0.1
aiofiles==23.2.1
